% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_model_assessment.R
\name{bias_smc_no_smc_month}
\alias{bias_smc_no_smc_month}
\title{Calculate and Compare Bias for SMC and Non-SMC Months}
\usage{
bias_smc_no_smc_month(obs_cases, simulated_df, smc_data)
}
\arguments{
\item{obs_cases}{A data frame containing observed cases with columns:
\itemize{
\item \code{date_ymd}: Date in "YYYY-MM-DD" format.
\item \code{month_no}: Month number.
\item \code{inc}: Total cases.
}}

\item{simulated_df}{A data frame containing simulated cases with columns:
\itemize{
\item \code{date_ymd}: Date in "YYYY-MM-DD" format.
\item \code{month_no}: Month number.
\item \code{inc}: Total simulated cases.
}}

\item{smc_data}{A data frame containing SMC information with columns:
\itemize{
\item \code{dates}: Date in "YYYY-MM-DD" format.
\item \code{SMC}: Binary indicator (1 for SMC applied, 0 otherwise).
}}
}
\value{
A data frame summarizing the bias in SMC and non-SMC months, containing:
\itemize{
\item \code{category}: "SMC Months" or "Non-SMC Months".
\item \code{avg_bias}: Average bias (observed - simulated).
\item \code{sd_bias}: Standard deviation of the bias.
\item \code{count}: Number of months in each category.
}
}
\description{
This function calculates the bias (difference between observed and simulated cases)
for malaria cases during months with and without Seasonal Malaria Chemoprevention (SMC).
It summarizes the average bias, standard deviation, and the number of months in
each category (SMC and non-SMC).
}
\examples{
result <- bias_smc_no_smc_month(obs_cases, simulated_df, smc_data)
print(result)

}
