% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_model_assessment.R
\name{calculate_bias}
\alias{calculate_bias}
\title{Calculate Bias}
\usage{
calculate_bias(observed, simulated)
}
\arguments{
\item{observed}{A numeric vector of observed values.}

\item{simulated}{A numeric vector of simulated values.}
}
\value{
The bias value.
}
\description{
Calculates the bias, which shows if the model tends to underpredict or overpredict.
Positive bias means overprediction, while negative bias means underprediction.
}
