% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_model_assessment.R
\name{calculate_bias_per_year}
\alias{calculate_bias_per_year}
\title{Calculate Bias Per Year}
\usage{
calculate_bias_per_year(simulated_df, obs_cases)
}
\arguments{
\item{simulated_df}{A data frame containing the simulated data with a \code{date_ymd} column and metrics such as \code{inc_A}, \code{inc_C}, and \code{inc}.}

\item{obs_cases}{A data frame containing the observed data with a \code{date_ymd} column and metrics such as \code{inc_A}, \code{inc_C}, and \code{inc}.}
}
\value{
A data frame summarizing the total bias for each metric (\code{inc_A}, \code{inc_C}, \code{inc}) per year.
The output includes the columns:
\itemize{
\item \code{year}: The year.
\item \code{total_bias_inc_A}: Total bias for \code{inc_A} in the year.
\item \code{total_bias_inc_C}: Total bias for \code{inc_C} in the year.
\item \code{total_bias_inc}: Total bias for \code{inc} in the year.
}
}
\description{
This function calculates the bias between simulated and observed data for each year.
}
\examples{
simulated_df <- data.frame(
  date_ymd = as.Date(c("2018-01-01", "2018-01-02", "2019-01-01")),
  inc_A = c(5, 3, 4),
  inc_C = c(2, 1, 3),
  inc = c(7, 4, 7)
)
obs_cases <- data.frame(
  date_ymd = as.Date(c("2018-01-01", "2018-01-02", "2019-01-01")),
  inc_A = c(6, 4, 5),
  inc_C = c(3, 2, 4),
  inc = c(9, 6, 9)
)
calculate_bias_per_year(simulated_df, obs_cases)

}
