% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterfactual_analysis.R
\name{calculate_estimate}
\alias{calculate_estimate}
\title{Calculate Outcome Across Matched Simulation Lists}
\usage{
calculate_estimate(o1, o2, outcome_fn)
}
\arguments{
\item{o1}{List of data frames (e.g., with SMC).}

\item{o2}{List of data frames (e.g., no SMC).}

\item{outcome_fn}{Function that takes two data frames and returns a scalar outcome.}
}
\value{
Numeric vector of outcomes (e.g., effectiveness or cases averted) for each parameter set.
}
\description{
Calculate Outcome Across Matched Simulation Lists
}
