% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_simulation.R
\name{calculate_incidence_quantiles}
\alias{calculate_incidence_quantiles}
\title{Calculate Quantiles for Incidence from Multiple Simulations}
\usage{
calculate_incidence_quantiles(simulations)
}
\arguments{
\item{simulations}{A list of data frames, each containing the simulation output for different parameter sets.}
}
\value{
A data frame containing the quantiles for each group (\code{inc_A}, \code{inc_C}, and \code{inc}) for each time point.
}
\description{
This function calculates the 1st and 99th quantiles of the simulated incidence data across multiple simulations,
for each group (\code{inc_A}, \code{inc_C}, and \code{inc}).
}
\examples{
# Assuming `simulations` is a list of data frames with simulated incidence data
incidence_quantiles <- calculate_incidence_quantiles(simulations)
}
