% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_model_assessment.R
\name{calculate_mae}
\alias{calculate_mae}
\title{Calculate Mean Absolute Error (MAE)}
\usage{
calculate_mae(observed, simulated)
}
\arguments{
\item{observed}{A numeric vector of observed values.}

\item{simulated}{A numeric vector of simulated values.}
}
\value{
The MAE value.
}
\description{
Calculates the Mean Absolute Error between observed and simulated values.
}
\examples{
mae_value <- calculate_mae(observed, simulated)
}
