% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_model_assessment.R
\name{calculate_rmse}
\alias{calculate_rmse}
\title{Calculate Root Mean Squared Error (RMSE)}
\usage{
calculate_rmse(observed, simulated)
}
\arguments{
\item{observed}{A numeric vector of observed values.}

\item{simulated}{A numeric vector of simulated values.}
}
\value{
The RMSE value.
}
\description{
Calculates the Root Mean Squared Error between observed and simulated values.
}
\examples{
rmse_value <- calculate_rmse(observed, simulated)
}
