% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_setup.R
\name{create_mcmc_params}
\alias{create_mcmc_params}
\title{Create Adaptive Proposal and MCMC Control Parameters}
\usage{
create_mcmc_params(
  stage = "stage1",
  initial_vcv_weight = 1,
  initial_scaling = 2,
  initial_scaling_weight = NULL,
  min_scaling = 0.1,
  scaling_increment = NULL,
  log_scaling_update = TRUE,
  acceptance_target = 0.234,
  forget_rate = 0.6,
  forget_end = Inf,
  adapt_end = Inf,
  pre_diminish = 40000,
  n_steps = 10000,
  n_burnin = 0,
  n_chains = 4,
  n_workers = 4,
  n_threads_total = 8
)
}
\arguments{
\item{stage}{possible values are "NULL", stage1", "stage2", or "noadapt"}

\item{initial_vcv_weight}{Weight for the initial variance-covariance matrix (default = 1).}

\item{initial_scaling}{Scaling factor for the proposal (default = 2).}

\item{initial_scaling_weight}{Optional weight for initial scaling (default = NULL).}

\item{min_scaling}{Minimum scaling factor for the proposal (default = 0.1).}

\item{scaling_increment}{Increment for scaling factor (default = NULL).}

\item{log_scaling_update}{Logical, should the scaling be updated on a log scale? (default = TRUE).}

\item{acceptance_target}{The target acceptance rate for the proposal (default = 0.234).}

\item{forget_rate}{Rate at which the proposal forgets past history (default = 0.6).}

\item{forget_end}{Time step at which the forgetting stops (default = Inf).}

\item{adapt_end}{Time step at which the adaptation stops (default = Inf).}

\item{pre_diminish}{Time steps before the diminishing starts (default = 40000).}

\item{n_steps}{Total number of MCMC steps (default = 10000).}

\item{n_burnin}{Number of burn-in steps (default = 0).}

\item{n_chains}{Number of MCMC chains (default = 4).}

\item{n_workers}{Number of workers for parallel execution (default = 4).}

\item{n_threads_total}{Total number of threads for parallel execution (default = 8).}
}
\value{
A list containing the adaptive proposal parameters and MCMC control parameters.
}
\description{
This function defines the default values for adaptive proposal control parameters
and MCMC control parameters, while allowing users to specify custom values.
}
