% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_aux_functions.R
\name{create_proposal_matrix}
\alias{create_proposal_matrix}
\title{Create a Proposal Matrix}
\usage{
create_proposal_matrix(proposal_variance, params_to_estimate, paramList_name)
}
\arguments{
\item{proposal_variance}{A named list or vector containing the variance for each parameter.}

\item{params_to_estimate}{A character vector of parameter names to estimate.}

\item{paramList_name}{A character vector of all parameter names.}
}
\value{
A square matrix with dimensions equal to the number of parameters, with the diagonal elements corresponding to the proposal variances for each parameter.
}
\description{
This function generates a proposal matrix for use in MCMC methods, where each parameter has its own proposal variance.
}
\examples{
proposal_matrix <- create_proposal_matrix(proposal_variance = list(phi = 0.01, qR = 0.1), params_to_estimate = c("phi", "qR"), paramList_name = c("phi", "qR", "mu_IR"))
}
