% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_aux_functions.R
\name{create_start_values}
\alias{create_start_values}
\title{Create Starting Values for Parameters}
\usage{
create_start_values(
  params_to_estimate,
  control_params,
  paramList_name,
  random = TRUE,
  seed = 10
)
}
\arguments{
\item{params_to_estimate}{A character vector of parameter names that are to be estimated.}

\item{control_params}{A list of control parameters (must include \code{n_chains} to specify the number of chains).}

\item{paramList_name}{A character vector of all parameter names.}

\item{random}{A logical value. If \code{TRUE}, random values are drawn from a uniform distribution; if \code{FALSE}, values are evenly spaced between the lower and upper bounds.}

\item{seed}{An integer specifying the seed for random number generation (default is 10).}
}
\value{
A matrix of starting values for each parameter and chain.
}
\description{
This function generates starting values for parameters based on the specified method. It can either draw random values from a uniform distribution or assign values from a specified interval.
}
\examples{
start_values <- create_start_values(params_to_estimate = c("phi", "qR"), control_params = list(n_chains = 4), paramList_name = c("phi", "qR", "mu_IR"))
}
