% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_simulation.R
\name{data_sim}
\alias{data_sim}
\title{Simulate Data from Model}
\usage{
data_sim(
  model,
  param_inputs,
  start_date,
  end_date,
  month = FALSE,
  round = TRUE,
  save = TRUE,
  file = "",
  month_unequal_days = FALSE,
  return_EIR = FALSE
)
}
\arguments{
\item{model}{The model to be simulated.}

\item{param_inputs}{A list of parameters required for the simulation.}

\item{start_date}{The start date for the simulation, as a \code{Date} object or character string.}

\item{end_date}{The end date for the simulation, as a \code{Date} object or character string.}

\item{month}{Logical; if \code{TRUE}, aggregate data monthly; if \code{FALSE}, aggregate weekly.}

\item{round}{Logical; if \code{TRUE}, round the results to whole numbers.}

\item{save}{Logical; if \code{TRUE}, save the results to a file.}

\item{file}{Character; file name for saving the results, if \code{save} is \code{TRUE}.}

\item{month_unequal_days}{Logical; if \code{TRUE}, adjust aggregation for unequal days in months.}
}
\value{
A data frame containing the incidence data aggregated by month or week.
}
\description{
This function runs the simulation using the specified model and parameters,
and returns either monthly or weekly aggregated incidence data for adults (\code{inc_A})
and children (\code{inc_C}), as well as total incidence (\code{inc}).
}
\examples{
# Running the simulation and getting monthly data
inc_data <- data_sim(model, param_inputs, start_date = "2021-01-01", end_date = "2021-12-31", month = TRUE)
}
