% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{date_to_months}
\alias{date_to_months}
\title{Convert Date Range to Monthly Format}
\usage{
date_to_months(start_date, end_date)
}
\arguments{
\item{start_date}{The start date as a character string or \code{Date} object.}

\item{end_date}{The end date as a character string or \code{Date} object.}
}
\value{
A vector of \code{Date} objects representing the first day of each month within the given date range.
}
\description{
This function takes a start date and an end date, and converts the range into a vector of the first day
of each month in the range, formatted as "YYYY-MM".
}
\examples{
date_to_months("2022-01-01", "2022-06-01")
}
