% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{date_to_weeks}
\alias{date_to_weeks}
\title{Convert a Date Range to Weekly Year Format}
\usage{
date_to_weeks(start_date, end_date)
}
\arguments{
\item{start_date}{The start date as a character string or \code{Date} object.}

\item{end_date}{The end date as a character string or \code{Date} object.}
}
\value{
A character vector containing each week between \code{start_date} and \code{end_date} in the format "YYYY-WW".
}
\description{
This function takes a start date and an end date, and converts the range into
a vector of week-year format ("YYYY-WW") according to the ISO week numbering.
}
\examples{
date_to_weeks("2022-01-01", "2022-03-01")
}
