% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_simulation.R
\name{decay_SMC}
\alias{decay_SMC}
\title{Compute SMC Efficacy Decay Over Time}
\usage{
decay_SMC(x, const = -0.1806)
}
\arguments{
\item{x}{Numeric. Number of days since the last SMC treatment.}

\item{const}{Numeric. The decay constant that controls the steepness of the sigmoid curve. Default is -0.1806.}
}
\value{
Numeric. Remaining efficacy of SMC treatment at day \code{x}.
}
\description{
Models the decay of Seasonal Malaria Chemoprevention (SMC) efficacy as a sigmoid function
of time since last administration.
}
\examples{
decay_SMC(0)      # Should return near maximum efficacy
decay_SMC(60)     # Should return reduced efficacy

}
