% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_processing_functions.R
\name{extract_era5}
\alias{extract_era5}
\title{Extract Temperature Data from ERA5 File}
\usage{
extract_era5(lon, lat, path_to_file)
}
\arguments{
\item{lon}{A numeric value specifying the longitude of the location.}

\item{lat}{A numeric value specifying the latitude of the location.}

\item{path_to_file}{A string specifying the path to the NetCDF file containing ERA5 data.}
}
\value{
A dataframe containing two columns: \code{Date} (POSIXct) and \code{Temperature} (in Celsius).
}
\description{
Extracts temperature data for a specified longitude and latitude from a NetCDF ERA5 file.
}
\examples{
\dontrun{
extract_era5(lon = 10.5, lat = -25.3, path_to_file = "./data/era5_011220.nc")
}
}
