% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_functions.R
\name{get_observation_function}
\alias{get_observation_function}
\title{Get Observation Function Based on User Configuration}
\usage{
get_observation_function(
  month = TRUE,
  age_for_inf = "u5",
  incidence_df,
  include_prev = TRUE,
  use_SMC_as_covariate = FALSE
)
}
\arguments{
\item{month}{Logical. TRUE for monthly data, FALSE for weekly.}

\item{age_for_inf}{Character. One of: "total", "sep_ages", "u5", "o5", "all_ages".}

\item{incidence_df}{A data frame of observed incidence and prevalence values.}

\item{include_prev}{Logical. If TRUE, includes prevalence likelihood.}

\item{use_SMC_as_covariate}{Logical. If TRUE, uses \code{beta_1} covariate effect on incidence.}
}
\value{
A likelihood function that takes (state, observed, pars) as arguments.
}
\description{
This helper function allows users to specify key options such as time resolution,
age group, inclusion of prevalence, and whether to model SMC as a covariate using \code{beta_1}.
}
