% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_simulation.R
\name{import_model}
\alias{import_model}
\title{Import an Odin Model into the malclimsim Package}
\usage{
import_model(model_path, model_name)
}
\arguments{
\item{model_path}{Full file path to the external Odin model (an R file) to be imported.}

\item{model_name}{Desired name (without ".R") for the model when stored inside the package.}
}
\value{
None. The function copies the model file into the package's models directory.
}
\description{
Import an Odin Model into the malclimsim Package
}
\examples{
import_model("path/to/external_model.R", "my_new_model")
}
