% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_scheduling.R
\name{load_clean_smc_data}
\alias{load_clean_smc_data}
\title{Load and clean raw SMC coverage data}
\usage{
load_clean_smc_data(data = NULL, path_to_excel = NULL, path_to_rds = NULL)
}
\arguments{
\item{data}{A data frame containing raw SMC coverage data (default = NULL).}

\item{path_to_excel}{Optional. Path to an Excel file containing SMC data (default = NULL).}

\item{path_to_rds}{Optional. Path to an RDS file containing SMC data (default = NULL).}
}
\value{
A cleaned data frame with columns: \code{date_start} and \code{coverage}.
Only one row per Year-Month is retained (the first entry chronologically).
}
\description{
Cleans and formats SMC coverage data. Allows for flexibility in input:
the user can supply either a data frame directly or a path to a file
(Excel or RDS). If both \code{data} and paths are \code{NULL}, the function returns an error.
}
