% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_simulation.R
\name{load_model}
\alias{load_model}
\title{Loading an Odin Model}
\usage{
load_model(name)
}
\arguments{
\item{name}{name of the model written in Odin DSL to be loaded. This is a path to an R file,}
}
\value{
the loaded model to be used for simulation, inference, etc
}
\description{
Loading an Odin Model
}
