% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_aux_functions.R
\name{make_transform}
\alias{make_transform}
\title{Create a Parameter Transformation Function}
\usage{
make_transform(temp, c_R_D, SMC, decay, cov_SMC)
}
\arguments{
\item{temp}{A numeric value representing the temperature parameter.}

\item{c_R_D}{A numeric value representing the decay rate.}

\item{SMC}{A numeric value representing the SMC (Scale of Mass Concentration).}

\item{decay}{A numeric value representing the decay parameter.}

\item{cov_SMC}{A numeric value representing the covariance for SMC.}
}
\value{
A function that takes a vector of parameters (\code{theta}) and combines them with the other provided values into a list.
}
\description{
This function returns a transformation function that appends parameter values to a list. It is used for transforming model parameters into a format compatible with other functions.
}
\examples{
transform_function <- make_transform(temp = 25, c_R_D = 0.5, SMC = 0.1, decay = 0.03, cov_SMC = 0.05)
transformed_params <- transform_function(c(0.1, 0.2, 0.3))
}
