% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_diagnostics.R
\name{max_ll_post}
\alias{max_ll_post}
\title{Extract and Print Maximum Log-Likelihood and Log-Posterior}
\usage{
max_ll_post(results, print_ll = TRUE)
}
\arguments{
\item{results}{A list containing the MCMC results from an inference run,
including the \code{coda_pars} matrix with columns for \code{log_likelihood} and \code{log_posterior}.}

\item{print_ll}{Boolean. If \code{TRUE}, the maximum log-likelihood and log-posterior values
will be printed to the console (default: \code{TRUE}).}
}
\value{
A list with two named elements:
\describe{
\item{max_log_likelihood}{The maximum value of the log-likelihood.}
\item{max_log_posterior}{The maximum value of the log-posterior.}
}
}
\description{
This function extracts and optionally prints the maximum values of the log-likelihood
and log-posterior from the results of an MCMC run. It is useful for quickly evaluating
the fit of a model.
}
