% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obs_cases}
\alias{obs_cases}
\title{Observed Malaria Incidence Used for Model Calibration}
\format{
A data frame with the following columns:
\describe{
\item{date_ymd}{Date corresponding to each week (class: Date)}
\item{inc_A}{Simulated incidence in adults}
\item{inc_C}{Simulated incidence in children}
\item{week_no}{Week index relative to simulation start}
}
}
\source{
Generated using \code{data_sim()} with MAP-calibrated model parameters.
}
\usage{
obs_cases
}
\description{
Weekly simulated malaria incidence data used to represent observed cases for calibration
and validation of the transmission model. Generated deterministically using maximum a posteriori (MAP) parameter values.
}
\keyword{datasets}
