% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_diagnostics.R
\name{plot_chains}
\alias{plot_chains}
\title{Convert MCMC Output to List of Chains for Visualization}
\usage{
plot_chains(mcmc_run)
}
\arguments{
\item{mcmc_run}{An object returned by \code{mcstate::pmcmc()}, containing posterior samples and a \code{chain} column indicating chain IDs.}
}
\value{
An \code{mcmc.list} object (from the \code{coda} package), where each element is an individual chain converted to an \code{mcmc} object.
}
\description{
Converts the output of a multi-chain MCMC run into a \code{coda::mcmc.list} object, separating chains
for easier visualization and diagnostics (e.g., trace plots or Gelman-Rubin diagnostics).
}
\details{
This function assumes that \code{mcmc_run$pars} contains the posterior samples and that there is a
\code{chain} vector in \code{mcmc_run$chain} specifying the chain each row belongs to.
}
\examples{
\dontrun{
mcmc_chains <- plot_chains(mcmc_run)
plot(mcmc_chains)  # Trace plots for all chains
}

}
