% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_diagnostics.R
\name{plot_corr}
\alias{plot_corr}
\title{Plot Correlation of MCMC Samples}
\usage{
plot_corr(results, title = NULL)
}
\arguments{
\item{results}{list: Results from the \code{inf_run} function, containing MCMC samples.}

\item{title}{character: Optional title for the plot.}
}
\value{
A \code{GGally} ggpairs object visualizing the correlation structure of MCMC samples.
}
\description{
This function creates a correlation plot for MCMC samples, showing the relationships between sampled parameters.
The upper panels display correlation values, and the lower panels show smoothed plots of parameter relationships.
}
