% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_distributions.R
\name{plot_priors}
\alias{plot_priors}
\title{Plot Default Priors}
\usage{
plot_priors(param_inputs, proposal_matrix, params_to_estimate, priors = NULL)
}
\arguments{
\item{param_inputs}{List of input parameters for the model.}

\item{proposal_matrix}{Covariance matrix for the MCMC proposal distribution.}

\item{params_to_estimate}{Character vector specifying which parameters' priors should be plotted.}

\item{priors}{List of priors to plot. If NULL, the default priors will be displayed.}
}
\value{
A ggplot object showing the prior distributions for each of the specified parameters.
}
\description{
This function plots the default priors for the specified parameters, showing their distributions.
}
\examples{
plot_priors(param_inputs, proposal_matrix, params_to_estimate = c("a_R", "b_R", "qR", "z", "eff_SMC", "phi", "size"))
}
