% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_functions.R
\name{print_observation_structure}
\alias{print_observation_structure}
\title{Print Configuration for Observation Function}
\usage{
print_observation_structure(
  month,
  age_for_inf,
  include_prev,
  use_beta_adjustment = FALSE
)
}
\description{
Utility to confirm current likelihood settings (time unit, age group, etc.).
}
