% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_aux_functions.R
\name{relate_time_steps_to_observed_data}
\alias{relate_time_steps_to_observed_data}
\title{Relate Time Steps to Observed Data for mcstate}
\usage{
relate_time_steps_to_observed_data(
  incidence_observed,
  month = FALSE,
  initial_time_obs = 0,
  rate = 7
)
}
\arguments{
\item{incidence_observed}{A data frame containing observed incidence data (cumulative incidence or other types of observations).}

\item{month}{A logical value indicating whether the data is provided in monthly intervals (\code{TRUE}) or weekly intervals (\code{FALSE}).}

\item{initial_time_obs}{An integer specifying the initial time step for monthly data (defaults to 0 for weekly data).}

\item{rate}{An integer specifying the rate (e.g., 7 for weekly data, 30 for monthly data) for resampling the time steps.}
}
\value{
A data frame with filtered data, including time intervals and the appropriate columns for particle filtering.
}
\description{
This function formats observed data into a form usable by the \code{mcstate} particle filter, adjusting for time steps (weeks or months) and ensuring the data is in the correct format for particle filtering.
}
\examples{
# Assuming 'incidence_observed' is a data frame with your observed data
formatted_data <- relate_time_steps_to_observed_data(incidence_observed, month = FALSE, initial_time_obs = 0, rate = 7)

}
