% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_distributions.R
\name{return_default_priors}
\alias{return_default_priors}
\title{Return Default Priors Used for Inference Procedure}
\usage{
return_default_priors()
}
\description{
Return Default Priors Used for Inference Procedure
}
