% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_processing_functions.R
\name{save_climate_data}
\alias{save_climate_data}
\title{Saving Climate Data From ERA5 and CHIRTSdaily}
\usage{
save_climate_data(
  lon,
  lat,
  years,
  path_to_data,
  rain = TRUE,
  temp = TRUE,
  temp_file_name = NULL,
  rain_file_name = NULL
)
}
\arguments{
\item{lon}{\itemize{
\item longitude coordinate
}}

\item{lat}{\itemize{
\item latitude coordinate
}}

\item{years}{\itemize{
\item vector containing each year for data to be downloaded
}}

\item{path_to_data}{\itemize{
\item folder where data is to be saved
}}

\item{rain}{\itemize{
\item boolean for if CHIRPS data should be downloaded and saved
}}

\item{temp}{\itemize{
\item boolean for if ERA5 data should be downloaded and saved
}}

\item{temp_file_name}{\itemize{
\item character file name for stored temperature data (ERA5)
}}

\item{rain_file_name}{\itemize{
\item character file name for stored rainfall data (CHIRPS)
}}
}
\value{
\itemize{
\item nothing
}
}
\description{
Saving Climate Data From ERA5 and CHIRTSdaily
}
\examples{
\dontrun{
years <- 2014:2023
lon <- 17.9
lat <- 8.3
path_to_data <- "C:/Users/putnni/Documents/r-packages/data/"
save_climate_data(lon, lat, years, path_to_data)
}
}
