% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_processing_functions.R
\name{save_era5}
\alias{save_era5}
\title{Save ERA5 Reanalysis Data}
\usage{
save_era5(years, path, file_name = NULL)
}
\arguments{
\item{years}{A vector of years for which the data is to be downloaded (e.g., \code{c(2000, 2001)}).}

\item{path}{A string specifying the path where the downloaded data file should be stored.}
}
\value{
The path to the downloaded NetCDF file.
}
\description{
Downloads ERA5 reanalysis data for specified years and saves it to the specified path.
}
\examples{
\dontrun{
save_era5(years = c(2000, 2001), path = "./data")
}
}
