% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterfactual_analysis.R
\name{set_eff_smc_to_zero}
\alias{set_eff_smc_to_zero}
\title{Set eff_SMC Parameter to Zero}
\usage{
set_eff_smc_to_zero(param_inputs)
}
\arguments{
\item{param_inputs}{A list of parameter values for the model.}
}
\value{
A list with the \code{eff_SMC} parameter set to zero.
}
\description{
This function takes a list of model parameters and sets the \code{eff_SMC} parameter to zero to simulate
a scenario without Seasonal Malaria Chemoprevention (SMC).
}
\examples{
no_smc_params <- set_eff_smc_to_zero(param_inputs)
}
