% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterfactual_analysis.R
\name{set_smc_coverage}
\alias{set_smc_coverage}
\title{Set SMC Coverage Level}
\usage{
set_smc_coverage(param_inputs, coverage_level = 1)
}
\arguments{
\item{param_inputs}{A list of parameter values for the model.}

\item{coverage_level}{A numeric value (between 0 and 1) representing the desired coverage level. Default is 1 (100\% coverage).}
}
\value{
A list with the \code{cov_SMC} parameter updated to the specified coverage level.
}
\description{
This function takes a list of model parameters and sets the \code{cov_SMC} parameter to a user-defined level.
}
\examples{
full_coverage_params <- set_smc_coverage(param_inputs, coverage_level = 1)
}
