% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_simulation.R
\name{sim_mod}
\alias{sim_mod}
\title{Simulate the Malaria Model Over Time}
\usage{
sim_mod(odin_mod, pars, time_start, n_particles, sim_time)
}
\arguments{
\item{odin_mod}{An \code{odin} model object (usually created with \code{odin::odin()}), representing the compiled malaria model.}

\item{pars}{A named list of parameters to pass to the model.}

\item{time_start}{Numeric value specifying the starting time (e.g., 0).}

\item{n_particles}{Integer; number of particles used in the stochastic simulation (i.e., number of replicate simulations).}

\item{sim_time}{Integer; total number of time steps (days) to simulate.}
}
\value{
A list with two elements:
\describe{
\item{x}{A 3D array of simulation results with dimensions \verb{[state, particle, time]}.}
\item{model}{The \code{odin} model object after simulation (can be used to inspect internals).}
}
}
\description{
Runs a stochastic malaria transmission model using an \code{odin}-generated model object, storing state outputs across multiple time steps and particles.
}
\details{
The model is initialized once using the \code{odin_mod$new()} constructor, and then \code{model$run(t)} is called at each time step \code{t} to advance the simulation.
}
\examples{
\dontrun{
mod <- odin::odin("malaria_model.R")
results <- sim_mod(mod, pars = list(beta = 0.2), time_start = 0, n_particles = 10, sim_time = 100)
}

}
