% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_simulation.R
\name{simulate_best}
\alias{simulate_best}
\title{Run the model once at the maximum-posterior parameter set (deterministic)}
\usage{
simulate_best(model, param_inputs, best_params, start_date, end_date, ...)
}
\arguments{
\item{model}{A dust or odin model object.}

\item{param_inputs}{Named list of baseline inputs.}

\item{best_params}{Named vector or list of best-fit parameters.}

\item{start_date}{Date or string start of sim window.}

\item{end_date}{Date or string end of sim window.}

\item{...}{Additional args passed to \code{data_sim()}.}
}
\value{
A data.frame in long form with \code{date_ymd}, \code{variable}, \code{value}.
}
\description{
Run the model once at the maximum-posterior parameter set (deterministic)
}
