% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterfactual_analysis.R
\name{simulate_with_smc}
\alias{simulate_with_smc}
\title{Simulate Model with SMC}
\usage{
simulate_with_smc(
  results,
  start_date,
  end_date,
  model,
  mu_transform_C = NULL,
  mu_transform_A = NULL,
  covariate_matrix = NULL
)
}
\arguments{
\item{results}{The MCMC results object containing the parameter samples.}

\item{start_date}{The start date for the simulation.}

\item{end_date}{The end date for the simulation.}

\item{model}{The model function to simulate from.}
}
\value{
A data frame containing the simulation output with SMC.
}
\description{
This function simulates the model using the parameters that maximize the log posterior, which include SMC.
}
