% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_distributions.R
\name{update_priors}
\alias{update_priors}
\title{Update Default Priors}
\usage{
update_priors(
  param_inputs,
  proposal_matrix,
  params_to_estimate,
  priors = NULL,
  new_priors
)
}
\arguments{
\item{priors}{List of priors to update. Default is NULL, which means using the default priors.}

\item{new_priors}{A named list of new prior specifications to replace the defaults.}
}
\value{
A list containing updated priors.
}
\description{
This function allows the user to update the default priors with new values.
}
