% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{climate_to_30_day_months}
\alias{climate_to_30_day_months}
\title{Convert Climate Data to 360-Day Years}
\usage{
climate_to_30_day_months(clim_df, start_year = 2014, end_year = 2022)
}
\arguments{
\item{clim_df}{A data frame of climate data, which must include columns for \code{date}, \code{month}, \code{day}, \code{anom}, \code{temp}, and \code{rollmean}.}

\item{start_year}{The starting year of the conversion (default is 2014).}

\item{end_year}{The ending year of the conversion (default is 2022).}
}
\value{
A data frame with 360-day years, where each month has exactly 30 days.
}
\description{
This function converts climate data to a 360-day calendar, with each month forced to 30 days.
Data is either removed or repeated to ensure the adjustment, useful for simplifying climate simulations.
}
\examples{
climate_to_30_day_months(clim_df, start_year = 2014, end_year = 2022)
}
