% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_setup.R
\name{create_proposal_matrix}
\alias{create_proposal_matrix}
\title{Create a Proposal Matrix}
\usage{
create_proposal_matrix(
  params_to_estimate,
  proposal_variance = NULL,
  correlation_matrix = NULL,
  model,
  param_inputs
)
}
\arguments{
\item{params_to_estimate}{A character vector of parameter names to estimate.}

\item{proposal_variance}{A named list containing the variance for each parameter. If \code{NULL}, default values are used.}

\item{model}{A model object with a \code{param()} method that returns all parameter names.}

\item{param_inputs}{A list of parameter values passed to initialize the model, used to extract parameters.}
}
\value{
A square matrix with dimensions equal to the number of parameters, with diagonal elements corresponding to the proposal variances for each parameter.
}
\description{
This function generates a proposal matrix for use in MCMC methods, where each parameter has its own proposal variance.
}
