% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_processing_functions.R
\name{daily_smooth_rain}
\alias{daily_smooth_rain}
\title{Title}
\usage{
daily_smooth_rain(cum_rain)
}
\arguments{
\item{cum_rain}{dataframe with two columns and rows equal to the number of weeks or months in original dataset}
}
\description{
Title
}
