% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{date_to_weeks}
\alias{date_to_weeks}
\title{Convert Date Range to Weekly Format (Sundays)}
\usage{
date_to_weeks(start_date, end_date)
}
\arguments{
\item{start_date}{The start date as a character string or \code{Date} object.}

\item{end_date}{The end date as a character string or \code{Date} object.}
}
\value{
A vector of \code{Date} objects representing each Sunday between the dates.
}
\description{
This function takes a start date and an end date, and returns a vector of Dates
representing the Sunday of each week in the range.
}
\examples{
date_to_weeks("2022-01-01", "2022-03-01")
}
