% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_setup.R
\name{extract_vcv}
\alias{extract_vcv}
\title{Title}
\usage{
extract_vcv(
  results,
  S_prev,
  save = TRUE,
  param_names,
  file_proposal = "",
  file_start = ""
)
}
\arguments{
\item{results}{list: results of inf_run function}

\item{S_prev}{numeric: number of previous steps to include in calculation}

\item{save}{boolean: whether or not to save the results}

\item{param_names}{list: list of model parameters}

\item{file_proposal}{character: file path for proposal cov matrix}

\item{file_start}{character: file path for chain start values}
}
\description{
Title
}
