% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_scheduling.R
\name{gen_smc_schedule}
\alias{gen_smc_schedule}
\title{Generate an SMC Deployment Schedule and Calculate Efficacy Decay}
\usage{
gen_smc_schedule(
  start_date,
  end_date,
  years,
  months_active,
  months_30_days = FALSE,
  coverage = 0.9,
  smc_day_of_month = 1
)
}
\arguments{
\item{start_date}{A string specifying the start date of the SMC deployment (format: "YYYY-MM-DD").}

\item{end_date}{A string specifying the end date of the SMC deployment (format: "YYYY-MM-DD").}

\item{years}{A numeric vector of years for which SMC will be deployed.}

\item{months_active}{A matrix where each row corresponds to a year and columns represent months
(1 = active, 0 = inactive).}

\item{months_30_days}{A logical flag to simulate a 360-day calendar. Default is \code{FALSE}.}

\item{coverage}{A numeric value specifying the coverage rate of SMC. Default is \code{0.90}.}

\item{smc_day_of_month}{Determines which day of the month the SMC round begins.}
}
\value{
A dataframe with columns:
\item{dates}{The sequence of dates for the given period.}
\item{SMC}{Binary values indicating SMC deployment (1 = deployed, 0 = not deployed).}
\item{cov}{The coverage rate of SMC over the period.}
\item{decay}{The calculated efficacy decay over time.}
}
\description{
This function simulates an SMC (Seasonal Malaria Chemoprevention) schedule based on given parameters
and calculates efficacy decay over time.
}
\details{
The function generates a sequence of dates and assigns SMC deployments based on active months
provided for each year. It also calculates the decay in efficacy over time for the deployed SMC.
}
\examples{
# Example usage:
years <- 2023:2024
months_active <- matrix(c(1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0,  # Year 2023
                          1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 1, 0), # Year 2024
                        nrow = 2, byrow = TRUE)
start_date <- "2023-01-01"
end_date <- "2024-12-31"
schedule <- gen_smc_schedule(start_date, end_date, years, months_active, coverage = 0.85)
head(schedule)

}
