% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_aux_functions.R
\name{index}
\alias{index}
\title{Create Index Mapping for Model Variables}
\usage{
index(info)
}
\arguments{
\item{info}{A list containing index information, typically provided from a model setup.}
}
\value{
A list containing two elements, \code{run} and \code{state}, each with named indices for different model variables.
}
\description{
This function creates a list of indices for different model variables, organizing them into 'run' and 'state' categories based on the provided info.
}
\examples{
model_indices <- index(info)
}
