% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{load_clean_smc_data}
\alias{load_clean_smc_data}
\title{Load and clean raw SMC coverage data}
\usage{
load_clean_smc_data(path_to_SMC)
}
\arguments{
\item{path_to_SMC}{Path to Excel file containing SMC data}
}
\value{
A cleaned data frame with columns: date_start, coverage
}
\description{
Load and clean raw SMC coverage data
}
