% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{model_date_360}
\alias{model_date_360}
\title{Given a start-year and a 360-day offset, return (Year,Mon,Day) in 30‑day months}
\usage{
model_date_360(day_index, start_year)
}
\description{
Given a start-year and a 360-day offset, return (Year,Mon,Day) in 30‑day months
}
