% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_diagnostics.R
\name{plot_corr}
\alias{plot_corr}
\title{Title}
\usage{
plot_corr(results, title = NULL)
}
\arguments{
\item{results}{results from inf_run function}

\item{title}{title of plot}
}
\description{
Title
}
