% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_visualizations.R
\name{plot_observed_vs_simulated}
\alias{plot_observed_vs_simulated}
\title{Plot Observed vs Simulated Data with Quantiles Ribbon and Optional Climate Data Facet}
\usage{
plot_observed_vs_simulated(
  results,
  obs_cases,
  start_date,
  end_date,
  model,
  add_ribbon = FALSE,
  n_samples = 100,
  groups = c("inc_A", "inc_C", "inc"),
  met = NULL,
  climate_facet = FALSE,
  prewarm_years = 2,
  days_per_year = 360,
  plot_title = NULL
)
}
\arguments{
\item{results}{The MCMC results object containing the parameter samples.}

\item{obs_cases}{A data frame of observed cases with columns for \code{date_ymd} (date), \code{inc_A}, \code{inc_C}, and \code{inc}.}

\item{start_date}{The start date for the simulation output as a \code{Date} object or character string.}

\item{end_date}{The end date for the simulation as a \code{Date} object or character string.}

\item{model}{The model function to simulate from.}

\item{add_ribbon}{Logical; if TRUE, adds a ribbon to the plot representing the 1st and 99th quantiles of the incidence data from additional simulations.}

\item{n_samples}{The number of samples to draw from the MCMC results for the quantiles ribbon.}

\item{groups}{A character vector specifying which groups to include in the plot (\code{inc_A}, \code{inc_C}, \code{inc}).}

\item{met}{Optional climate data frame to be included as a separate facet.}

\item{climate_facet}{Logical; if TRUE, adds climate data as a separate facet in the plot.}

\item{prewarm_years}{Integer; the number of years to simulate before the \code{start_date} for stabilization (default is 2 years).}

\item{days_per_year}{Integer; the number of days per year in the simulation (default is 360).}
}
\value{
A ggplot object displaying observed data as points, simulated data as a line, and an optional ribbon for quantiles.
}
\description{
This function simulates data from a model using parameters that maximize the log posterior,
then plots the observed and simulated incidence data for comparison. Optionally, a ribbon
representing the 1st and 99th quantiles of additional simulations can be included to visualize uncertainty.
It also allows for optional inclusion of climate data as a separate facet.
}
\examples{
# Assuming `results` contains MCMC output and `obs_cases` is the observed cases data
plot_observed_vs_simulated(results, obs_cases, start_date = "2014-01-01",
                           end_date = "2014-12-31", model = data_sim,
                           add_ribbon = TRUE, n_samples = 100, groups = c("inc_A", "inc_C", "inc"),
                           met = met_data, climate_facet = TRUE, prewarm_years = 2)
}
