% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_distributions.R
\name{plot_priors}
\alias{plot_priors}
\title{Plot Default Priors}
\usage{
plot_priors(param_inputs, proposal_matrix, params_to_estimate, priors = NULL)
}
\arguments{
\item{param_inputs}{A list of model inputs (used only if default priors are initialized).}

\item{proposal_matrix}{A covariance matrix used to generate default priors if \code{priors} is not supplied.}

\item{params_to_estimate}{A character vector specifying which parameters to include in the plot.}

\item{priors}{A named list of priors. Each prior should be a list containing \code{min}, \code{max}, and a \code{prior} function
that returns the log-density for a given numeric input. If \code{NULL}, default priors will be computed.}
}
\value{
A \code{ggplot} object showing the prior distributions of the selected parameters.
}
\description{
Plots the default or user-specified prior distributions for the parameters of interest.
}
