% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_processing_functions.R
\name{process_climate_data}
\alias{process_climate_data}
\title{Title}
\usage{
process_climate_data(lon, lat, years, D, temp_path, rain_path, path_to_data)
}
\arguments{
\item{lon}{\itemize{
\item longitude coordinate
}}

\item{lat}{\itemize{
\item latitude coordinate
}}

\item{years}{\itemize{
\item vector containing each year for the analysis
}}

\item{D}{\itemize{
\item the number of previous days to take an average over for the rainfall rolling mean
}}

\item{temp_path}{\itemize{
\item path to temperature data saved by save_climate_data
}}

\item{rain_path}{path to rainfall data saved by save_climate_data}

\item{path_to_data}{path to directory where the processed climate data will be stored}
}
\value{
a data frame
}
\description{
Title
}
\examples{
years <- 2014:2023
lon <- 17.9
lat <- 8.3
path_to_data <- "C:/Users/putnni/Documents/r-packages/data/"
save_climate_data(lon, lat, years, path_to_data)
temp_path <- paste0(path_to_data, "era5_11051417.nc")
rain_path <- paste0(path_to_data, "chirps_11051418.rds")
met <- process_climate_data(lon, lat, years, D = 30, temp_path = temp_path,
rain_path = rain_path, path_to_data
}
