% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_simulation.R
\name{run_simulations_from_samples}
\alias{run_simulations_from_samples}
\title{Run Simulations Given Sampled Parameters}
\usage{
run_simulations_from_samples(
  model,
  param_inputs,
  param_samples,
  start_date,
  end_date,
  prewarm_years = 2,
  mu_transform_C = NULL,
  mu_transform_A = NULL,
  covariate_matrix = NULL
)
}
\arguments{
\item{model}{Simulation model function.}

\item{param_inputs}{Baseline parameter inputs.}

\item{param_samples}{Sampled parameter matrix.}

\item{start_date, }{end_date Simulation date range.}

\item{prewarm_years}{Years of prewarming.}

\item{days_per_year}{Days in simulation year.}
}
\value{
List of simulation data frames.
}
\description{
Run Simulations Given Sampled Parameters
}
