% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_simulation.R
\name{sample_mcmc_steps}
\alias{sample_mcmc_steps}
\title{Sample Parameter Sets from MCMC Results}
\usage{
sample_mcmc_steps(mcmc_results, num_samples)
}
\arguments{
\item{mcmc_results}{A matrix or data frame containing MCMC posterior samples of parameters.}

\item{num_samples}{An integer specifying the number of samples to draw.}
}
\value{
A matrix of sampled parameter sets.
}
\description{
This function randomly samples a specified number of parameter sets from the MCMC posterior distribution.
}
\examples{
sampled_params <- sample_mcmc_steps(MCMC_results_2_3$mcmc_run$pars, 100)
}
