% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_simulation.R
\name{simulate_models}
\alias{simulate_models}
\title{Simulate Models Using Sampled Parameters}
\usage{
simulate_models(
  model,
  param_inputs,
  sampled_params,
  start_date,
  end_date,
  prewarm_years = 2,
  days_per_year = 360
)
}
\arguments{
\item{model}{The model function to simulate from.}

\item{param_inputs}{The initial list of parameters to update for each simulation.}

\item{sampled_params}{A list of parameter sets to use for each simulation.}

\item{start_date}{The start date for the simulation.}

\item{end_date}{The end date for the simulation.}
}
\value{
A list of data frames, each containing the simulation output for a different parameter set.
}
\description{
This function runs \code{n} model simulations, each using a different set of parameters sampled
from the MCMC results.
}
\examples{
# Assuming `sampled_params` is a list of sampled parameter sets
simulations <- simulate_models(model = data_sim, param_inputs = param_inputs,
                               sampled_params = sampled_params,
                               start_date = "2021-01-01", end_date = "2021-12-31")
}
