% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_scheduling.R
\name{smc_schedule_from_data}
\alias{smc_schedule_from_data}
\title{Generate a schedule of coverage values for a given period with decay}
\usage{
smc_schedule_from_data(smc_cov, months_30_days, years, const = -0.1806)

smc_schedule_from_data(smc_cov, months_30_days, years, const = -0.1806)
}
\arguments{
\item{smc_cov}{A data frame containing the coverage data. It should have at least two columns:
\itemize{
\item \code{date_start}: The date of the coverage value.
\item \code{coverage}: The coverage value corresponding to that date.
}}

\item{months_30_days}{Logical; whether to generate dates in 30-day month format (defaults to \code{FALSE}).
If \code{TRUE}, the sequence of dates will follow a 360-day year with 12 months of 30 days each.}

\item{years}{A numeric vector containing the range of years to generate the schedule for (e.g., \code{c(2014, 2022)}).}

\item{const}{A numeric value representing the decay constant used in the decay calculation (default is \code{-0.1806}).}
}
\value{
A data frame containing the following columns:
\itemize{
\item \code{dates}: The sequence of dates from the beginning to the end of the time period.
\item \code{SMC}: A binary value indicating the presence (1) or absence (0) of coverage on each date.
\item \code{cov}: The coverage value for each date.
\item \code{decay}: The decay values calculated for each date based on the SMC and the decay constant.
}

A data frame containing the following columns:
\itemize{
\item \code{dates}: The sequence of dates from the beginning to the end of the time period.
\item \code{SMC}: A binary value indicating the presence (1) or absence (0) of coverage on each date.
\item \code{cov}: The coverage value for each date.
\item \code{decay}: The decay values calculated for each date based on the SMC and the decay constant.
}
}
\description{
This function generates a schedule of coverage values for the given time period,
filling missing values with previous non-zero coverage and applying a decay calculation.

This function generates a schedule of coverage values for the given time period,
filling missing values with previous non-zero coverage and applying a decay calculation.
}
\examples{
smc_schedule_from_data(smc_cov = smc_cov_data, months_30_days = FALSE, years = c(2014, 2022))
smc_schedule_from_data(smc_cov = smc_cov_data, months_30_days = FALSE, years = c(2014, 2022))
}
