% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_simulation.R
\name{update_param_list}
\alias{update_param_list}
\title{Update Parameters in List with Values from Extracted Vector}
\usage{
update_param_list(param_inputs, param_values)
}
\arguments{
\item{param_inputs}{A list of parameters where each element is a named parameter.}

\item{param_values}{A named vector of parameters, typically from the maximum
log posterior, to update values in \code{param_inputs}.}
}
\value{
An updated list of parameters with replaced values where matches exist.
}
\description{
This function updates values in a list of parameters (e.g., \code{param_inputs})
with values from a named vector of parameters obtained from an MCMC run.
}
\examples{
# Assuming `param_inputs` is a list of parameters and `params_at_max_posterior`
# is a named vector with values to update
updated_params <- update_param_list(param_inputs, params_at_max_posterior)
}
